Moved Favourites Features into their own component
